+++ x/ecocredit/base/keeper/query_class_info.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
+	regenerrors "github.com/regen-network/regen-ledger/errors"
 	types "github.com/regen-network/regen-ledger/x/ecocredit/base/types/v1"
 )
 
@@ -13,7 +13,7 @@ import (
 func (k Keeper) Class(ctx context.Context, request *types.QueryClassRequest) (*types.QueryClassResponse, error) {
 	class, err := k.stateStore.ClassTable().GetById(ctx, request.ClassId)
 	if err != nil {
-		return nil, sdkerrors.ErrInvalidRequest.Wrapf("could not get class with id %s: %s", request.ClassId, err.Error())
+		return nil, regenerrors.ErrNotFound.Wrapf("could not get class with id %s: %s", request.ClassId, err.Error())
 	}
 
 	admin := sdk.AccAddress(class.Admin)