+++ x/ecocredit/basket/types/v1/tx.pb.go
@@ -66,11 +66,10 @@ type MsgCreate struct {
 	// date_criteria is the date criteria for batches admitted to the basket.
 	// At most, only one of the fields in the date_criteria should be set.
 	DateCriteria *DateCriteria `protobuf:"bytes,8,opt,name=date_criteria,json=dateCriteria,proto3" json:"date_criteria,omitempty"`
-	// fee is the basket creation fee. A fee is not required if the list of fees
-	// in Params.basket_fee is empty. The provided fee must be one of the fees
-	// listed in Params.basket_fee. The provided amount can be greater than
-	// or equal to the listed amount but the basket creator will only be charged
-	// the listed amount (i.e. the minimum amount).
+	// fee is the basket creation fee. A fee is not required if no fee exists
+	// in the basket fee parameter. The fee must be greater than or equal to the
+	// fee param. The curator will be charged the amount specified in the fee
+	// parameter, even if a greater amount is provided.
 	//
 	// Note (Since Revision 1): Although this field supports a list of fees, the
 	// basket creator must provide no more than one fee (i.e. one Coin in a list
@@ -230,9 +229,7 @@ type MsgPut struct {
 	// basket_denom is the basket denom to add credits to.
 	BasketDenom string `protobuf:"bytes,2,opt,name=basket_denom,json=basketDenom,proto3" json:"basket_denom,omitempty"`
 	// credits are credits to add to the basket. If they do not match the basket's
-	// admission criteria the operation will fail. If there are any "dust" credits
-	// left over when converting credits to basket tokens, these credits will
-	// not be converted to basket tokens and instead remain with the owner.
+	// admission criteria, the operation will fail.
 	Credits []*BasketCredit `protobuf:"bytes,3,rep,name=credits,proto3" json:"credits,omitempty"`
 }
 
@@ -345,7 +342,7 @@ type MsgTake struct {
 	// amount is the integer number of basket tokens to convert into credits.
 	Amount string `protobuf:"bytes,3,opt,name=amount,proto3" json:"amount,omitempty"`
 	// retirement_location is the optional retirement jurisdiction for the
-	// credits which will be used only if retire_on_take is true for this basket.
+	// credits which will be used only if retire_on_take is true.
 	//
 	// Deprecated (Since Revision 1): This field will be removed in the next
 	// version in favor of retirement_jurisdiction. Only one of these need to be
@@ -353,10 +350,12 @@ type MsgTake struct {
 	RetirementLocation string `protobuf:"bytes,4,opt,name=retirement_location,json=retirementLocation,proto3" json:"retirement_location,omitempty"` // Deprecated: Do not use.
 	// retire_on_take is a boolean that dictates whether the ecocredits
 	// received in exchange for the basket tokens will be received as
-	// retired or tradable credits.
+	// retired or tradable credits. If the basket has disable_auto_retire set to
+	// false, retire_on_take MUST be set to true, and a retirement jurisdiction
+	// must be provided.
 	RetireOnTake bool `protobuf:"varint,5,opt,name=retire_on_take,json=retireOnTake,proto3" json:"retire_on_take,omitempty"`
 	// retirement_jurisdiction is the optional retirement jurisdiction for the
-	// credits which will be used only if retire_on_take is true for this basket.
+	// credits which will be used only if retire_on_take is true.
 	//
 	// Since Revision 1
 	RetirementJurisdiction string `protobuf:"bytes,6,opt,name=retirement_jurisdiction,json=retirementJurisdiction,proto3" json:"retirement_jurisdiction,omitempty"`
@@ -776,12 +775,36 @@ const _ = grpc.SupportPackageIsVersion4
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://godoc.org/google.golang.org/grpc#ClientConn.NewStream.
 type MsgClient interface {
-	// Create creates a bank denom which wraps credits.
+	// Create creates a basket that can hold different types of ecocredits that
+	// meet the basket's criteria. Upon depositing ecocredits into the basket,
+	// basket tokens are minted and sent to depositor using the Cosmos SDK Bank
+	// module. This allows basket tokens to be utilized within IBC. Basket tokens
+	// are fully fungible with other basket tokens from the same basket. The
+	// basket token denom is derived from the basket name, credit type
+	// abbreviation, and credit type precision (i.e. basket name "foo", credit
+	// type exponent 6, and credit type abbreviation "C" generates the denom
+	// eco.uC.foo). Baskets can limit credit acceptance criteria based on a
+	// combination of credit type, credit classes, and credit batch start date.
+	// Credits can be taken from the basket in exchange for basket tokens. Taken
+	// credits will be immediately retired, unless disable_auto_retire is set to
+	// true. When set to true, credits may be received in either a tradable or
+	// retired state, depending on the taker's request. If the basket fee
+	// governance parameter is set, a fee of equal or greater value must be
+	// provided in the request. Only the amount specified in the fee parameter
+	// will be charged, even if a greater value fee is provided. Fees from
+	// creating a basket are burned.
 	Create(ctx context.Context, in *MsgCreate, opts ...grpc.CallOption) (*MsgCreateResponse, error)
-	// Put puts credits into a basket in return for basket tokens.
+	// Put deposits credits into the basket from the holder's tradable balance in
+	// exchange for basket tokens. The amount of tokens received is calculated by
+	// the following formula: sum(credits_deposited) * 10^credit_type_exponent.
+	// The credits being deposited MUST adhere to the criteria of the basket.
 	Put(ctx context.Context, in *MsgPut, opts ...grpc.CallOption) (*MsgPutResponse, error)
-	// Take takes credits from a basket starting from the oldest
-	// credits first.
+	// Take exchanges basket tokens for credits from the specified basket. Credits
+	// are taken deterministically, ordered by oldest batch start date to the most
+	// recent batch start date. If the basket has disable_auto_retire set to
+	// false, both retirement_jurisdiction and retire_on_take must be set, and the
+	// taken credits will be retired immediately upon receipt. Otherwise, credits
+	// may be received as tradable or retired, based on the request.
 	Take(ctx context.Context, in *MsgTake, opts ...grpc.CallOption) (*MsgTakeResponse, error)
 	// UpdateBasketFee is a governance method that allows for updating the basket
 	// creation fee. If not set, the basket creation fee will be removed and no
@@ -789,7 +812,7 @@ type MsgClient interface {
 	//
 	// Since Revision 2
 	UpdateBasketFee(ctx context.Context, in *MsgUpdateBasketFee, opts ...grpc.CallOption) (*MsgUpdateBasketFeeResponse, error)
-	// UpdateCurator updates basket curator
+	// UpdateCurator updates basket curator.
 	//
 	// Since Revision 2
 	UpdateCurator(ctx context.Context, in *MsgUpdateCurator, opts ...grpc.CallOption) (*MsgUpdateCuratorResponse, error)
@@ -850,12 +873,36 @@ func (c *msgClient) UpdateCurator(ctx context.Context, in *MsgUpdateCurator, opt
 
 // MsgServer is the server API for Msg service.
 type MsgServer interface {
-	// Create creates a bank denom which wraps credits.
+	// Create creates a basket that can hold different types of ecocredits that
+	// meet the basket's criteria. Upon depositing ecocredits into the basket,
+	// basket tokens are minted and sent to depositor using the Cosmos SDK Bank
+	// module. This allows basket tokens to be utilized within IBC. Basket tokens
+	// are fully fungible with other basket tokens from the same basket. The
+	// basket token denom is derived from the basket name, credit type
+	// abbreviation, and credit type precision (i.e. basket name "foo", credit
+	// type exponent 6, and credit type abbreviation "C" generates the denom
+	// eco.uC.foo). Baskets can limit credit acceptance criteria based on a
+	// combination of credit type, credit classes, and credit batch start date.
+	// Credits can be taken from the basket in exchange for basket tokens. Taken
+	// credits will be immediately retired, unless disable_auto_retire is set to
+	// true. When set to true, credits may be received in either a tradable or
+	// retired state, depending on the taker's request. If the basket fee
+	// governance parameter is set, a fee of equal or greater value must be
+	// provided in the request. Only the amount specified in the fee parameter
+	// will be charged, even if a greater value fee is provided. Fees from
+	// creating a basket are burned.
 	Create(context.Context, *MsgCreate) (*MsgCreateResponse, error)
-	// Put puts credits into a basket in return for basket tokens.
+	// Put deposits credits into the basket from the holder's tradable balance in
+	// exchange for basket tokens. The amount of tokens received is calculated by
+	// the following formula: sum(credits_deposited) * 10^credit_type_exponent.
+	// The credits being deposited MUST adhere to the criteria of the basket.
 	Put(context.Context, *MsgPut) (*MsgPutResponse, error)
-	// Take takes credits from a basket starting from the oldest
-	// credits first.
+	// Take exchanges basket tokens for credits from the specified basket. Credits
+	// are taken deterministically, ordered by oldest batch start date to the most
+	// recent batch start date. If the basket has disable_auto_retire set to
+	// false, both retirement_jurisdiction and retire_on_take must be set, and the
+	// taken credits will be retired immediately upon receipt. Otherwise, credits
+	// may be received as tradable or retired, based on the request.
 	Take(context.Context, *MsgTake) (*MsgTakeResponse, error)
 	// UpdateBasketFee is a governance method that allows for updating the basket
 	// creation fee. If not set, the basket creation fee will be removed and no
@@ -863,7 +910,7 @@ type MsgServer interface {
 	//
 	// Since Revision 2
 	UpdateBasketFee(context.Context, *MsgUpdateBasketFee) (*MsgUpdateBasketFeeResponse, error)
-	// UpdateCurator updates basket curator
+	// UpdateCurator updates basket curator.
 	//
 	// Since Revision 2
 	UpdateCurator(context.Context, *MsgUpdateCurator) (*MsgUpdateCuratorResponse, error)